# Conflicts:
#	scripts/globals/bcnm.lua
#	src/map/battlefield.cpp
#	src/map/battlefield.h
#	src/map/battlefield_handler.cpp
#	src/map/lua/lua_baseentity.cpp
#	src/map/lua/lua_battlefield.cpp
#	src/map/lua/luautils.h
#	src/map/utils/battlefieldutils.h
#	src/map/zone_entities.cpp

+++ src/common/utils.cpp
@@ -156,12 +156,12 @@ Rotations of entities are saved in uint8s, which can only hold up to a value of 
 */
 float rotationToRadian(uint8 rotation)
 {
-    return (((float)rotation) / 256) * 2 * M_PI;
+    return (float)((rotation / 256.0f) * 2 * M_PI);
 }
 
 uint8 radianToRotation(float radian)
 {
-    return (radian / (2 * M_PI)) * 256;
+    return (uint8)((radian / (2 * M_PI)) * 256);
 }
 
 
@@ -203,9 +203,9 @@ position_t nearPosition(const position_t& A, float offset, float radian)
     float totalRadians = rotationToRadian(A.rotation) + radian;
     position_t B;
 
-    B.x = A.x + cosf(2 * M_PI - totalRadians) * offset;
+    B.x = A.x + cosf((float)(2 * M_PI - totalRadians)) * offset;
     B.y = A.y;
-    B.z = A.z + sinf(2 * M_PI - totalRadians) * offset;
+    B.z = A.z + sinf((float)(2 * M_PI - totalRadians)) * offset;
 
     B.rotation = A.rotation;
     B.moving = A.moving;
@@ -496,7 +496,7 @@ int8* DecodeStringLinkshell(int8* signature, int8* target)
     for(uint8 currChar = 0; currChar < dsp_min(20, (strlen((const char*)signature) * 8) / 6); ++currChar)
     {
         uint8 tempChar = '\0';
-        tempChar = unpackBitsLE((uint8*)signature, currChar * 6, 6);
+        tempChar = (uint8)unpackBitsLE((uint8*)signature, currChar * 6, 6);
         if(tempChar >= 1 && tempChar <= 26)
             tempChar = 'a' - 1 + tempChar;
         else if(tempChar >= 27 && tempChar <= 52)
@@ -554,7 +554,7 @@ int8* DecodeStringSignature(int8* signature, int8* target)
     for(uint8 currChar = 0; currChar < dsp_min(15, (strlen((const char*)signature) * 8) / 6); ++currChar)
     {
         uint8 tempChar = '\0';
-        tempChar = unpackBitsLE((uint8*)signature, currChar * 6, 6);
+        tempChar = (uint8)unpackBitsLE((uint8*)signature, currChar * 6, 6);
         if(tempChar >= 1 && tempChar <= 10)
             tempChar = '0' - 1 + tempChar;
         else if(tempChar >= 11 && tempChar <= 36)
