+++ src/login/lobby.cpp
@@ -216,7 +216,7 @@ int32 lobbydata_parse(int32 fd)
                 WBUFW(CharList, 68 + 32 + i * 140) = (uint16)Sql_GetIntData(SqlHandle, 12); // main;
                 WBUFW(CharList, 70 + 32 + i * 140) = (uint16)Sql_GetIntData(SqlHandle, 13); // sub;
 
-                WBUFB(CharList, 72 + 32 + i * 140) = zone;
+                WBUFB(CharList, 72 + 32 + i * 140) = (uint8)zone;
                 WBUFW(CharList, 78 + 32 + i * 140) = zone;
                 ///////////////////////////////////////////////////
                 ++i;
